backend/requirements.txt
fastapi
uvicorn[standard]
motor
boto3
pydantic-settings
python-multipart
python-dotenv
numpy
Pillow
tflite-runtime
boto3
